+++ src/viam/app/viam_client.py
@@ -9,11 +9,6 @@ from viam.app.app_client import AppClient
 from viam.app.data_client import DataClient
 from viam.app.ml_training_client import MLTrainingClient
 from viam.rpc.dial import DialOptions, _dial_app, _get_access_token
-from viam.proto.app.data import (
-    CaptureInterval,
-    Filter,
-    TagsFilter,
-)
 from viam.utils import datetime_to_timestamp
 
 LOGGER = logging.getLogger(__name__)
@@ -89,65 +84,3 @@ class ViamClient:
         LOGGER.debug("Closing gRPC channel to app.")
         self._channel.close()
         self._closed = True
-
-    @staticmethod
-    def create_filter(
-        component_name: Optional[str] = None,
-        component_type: Optional[str] = None,
-        method: Optional[str] = None,
-        robot_name: Optional[str] = None,
-        robot_id: Optional[str] = None,
-        part_name: Optional[str] = None,
-        part_id: Optional[str] = None,
-        location_ids: Optional[List[str]] = None,
-        organization_ids: Optional[List[str]] = None,
-        mime_type: Optional[List[str]] = None,
-        start_time: Optional[datetime] = None,
-        end_time: Optional[datetime] = None,
-        tags: Optional[List[str]] = None,
-        bbox_labels: Optional[List[str]] = None,
-    ) -> Filter:
-        """Create a `Filter`.
-
-        Args:
-            component_name (Optional[str]): Optional name of the component that captured the data being filtered (e.g., "left_motor").
-            component_type (Optional[str]): Optional type of the componenet that captured the data being filtered (e.g., "motor").
-            method (Optional[str]): Optional name of the method used to capture the data being filtered (e.g., "IsPowered").
-            robot_name (Optional[str]): Optional name of the robot associated with the data being filtered (e.g., "viam_rover_1").
-            robot_id (Optional[str]): Optional ID of the robot associated with the data being filtered.
-            part_name (Optional[str]): Optional name of the system part associated with the data being filtered (e.g., "viam_rover_1-main").
-            part_id (Optional[str]): Optional ID of the system part associated with the data being filtered.
-            location_ids (Optional[List[str]]): Optional list of location IDs associated with the data being filtered.
-            organization_ids (Optional[List[str]]): Optional list of organization IDs associated with the data being filtered.
-            mime_type (Optional[List[str]]): Optional mime type of data being filtered (e.g., "image/png").
-            start_time (Optional[datetime.datetime]): Optional start time of an interval to filter data by.
-            end_time (Optional[datetime.datetime]): Optional end time of an interval to filter data by.
-            tags (Optional[List[str]]): Optional list of tags attached to the data being filtered (e.g., ["test"]).
-            bbox_labels (Optional[List[str]]): Optional list of bounding box labels attached to the data being filtered (e.g., ["square",
-                "circle"]).
-
-        Returns:
-            viam.proto.app.data.Filter: The `Filter` object.
-        """
-        return Filter(
-            component_name=component_name if component_name else "",
-            component_type=component_type if component_type else "",
-            method=method if method else "",
-            robot_name=robot_name if robot_name else "",
-            robot_id=robot_id if robot_id else "",
-            part_name=part_name if part_name else "",
-            part_id=part_id if part_id else "",
-            location_ids=location_ids,
-            organization_ids=organization_ids,
-            mime_type=mime_type,
-            interval=(
-                CaptureInterval(
-                    start=datetime_to_timestamp(start_time),
-                    end=datetime_to_timestamp(end_time),
-                )
-            )
-            if start_time or end_time
-            else None,
-            tags_filter=TagsFilter(tags=tags),
-            bbox_labels=bbox_labels,
-        )
